requirements.txt
appdirs==1.4.3
boto3==1.4.6
botocore==1.6.7
dj-database-url==0.4.2
Django==1.11
django-appconf==1.0.2
django-imagekit==4.0.1
django-next-prev==1.0.1
django-storages==1.6.5
docutils==0.14
gunicorn==19.7.1
jmespath==0.9.3
olefile==0.44
packaging==16.8
pilkit==2.0
Pillow==4.1.1
psycopg2==2.7.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
s3transfer==0.1.10
six==1.10.0
virtualenv==15.1.0
whitenoise==3.3.0
